clean code logic
