+++ lips/scoring/__init__.py
@@ -9,11 +9,10 @@
 from __future__ import absolute_import
 
 from lips.scoring.scoring import Scoring
-from lips.scoring.powergrid_scoring import PowerGridScoring
 from lips.scoring.airfoil_powergrid_scoring import AirfoilPowerGridScoring
 from lips.scoring.ml4physim_powergrid_socring import ML4PhysimPowerGridScoring
 
 
 __all__ = [
-    "Scoring", "PowerGridScoring", "AirfoilPowerGridScoring", "ML4PhysimPowerGridScoring"
+    "Scoring", "AirfoilPowerGridScoring", "ML4PhysimPowerGridScoring"
 ]
